testing/pyproject.toml
[tool.poetry]
name = "invariant"
version = "0.0.1"
description = "Testing library for AI Agents by InvariantLabs"
authors = ["InvariantLabs <hello@invariantlabs.ai>"]
readme = "README.md"
packages = [{ include = "invariant" }]

[tool.poetry.dependencies]
python = ">= 3.10, <4.0"
pydantic = "^2.9.2"
requests = "^2.32.3"
nltk = "^3.9.1"
pytest = "^8.3.3"
openai = "^1.54.4"
Pillow = "^10.0.0"
beautifulsoup4 = "^4.12.3"
invariant-sdk = "^0.0.4"
diskcache = "^5.6.3"

[tool.poetry.dev-dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
openai-swarm = "^0.1.1"
langgraph = "^0.2.53"
langchain-openai = "^0.2.10"
langchain-community = "^0.3.9"
anthropic = "^0.40.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)s] %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

# top-level command 'invariant'
[tool.poetry.scripts]
invariant = "invariant.__main__:main"

[tool.ruff.lint]
select = ["D"]

[tool.ruff.lint.pydocstyle]
convention = "google"